Fixes for Mac.
Adjusted the project file to use /opt/local/include and /opt/local/lib since
that has the older, working version of FFMPEG. The one in /usr/local might work
to but I was getting build problems when linking...likely because of the
myriad different places where /usr or /opt have to bet set and perfectly
synched up.

Additionally had to ensure that SUPRESS_EXCESS_RENDERING was set in project file
since it is set explicitly in the Linux Makefile.