Fixes FlatTiler bug and calls to clEnqueueWriteBuffer().
FlatTiler::UpdateDisplay was aggressively using OpenMP. It was
causing weird problems when mallocing tiles, but only when rendering
with OpenCL. I just reduced the use of OpenMP to the calculations of
checksums.

There were several calls to clEnqueueWriteBuffer() that used an
offset that was not in bytes. Corrected those. They were in methods
that were never used.