No longer instantiate a CoefficientMatrix for every matrix.
This leads to a tremendous memory savings and a substantial speedup.
This required the moving of the getters/setters form
CoeffiicentMatrix to CoefficientPlanes.

Additionally moved several fields from 32-bit values to size_t to
avoid calculations that accidently demote to 32-bit.

Lastly, the optical flow distance average is now normallized to the
diagonal of the video so we can compare flow between different
resolution images.